Moving ArtifactDescriptor to app-fabric
