+++ cdap-api/src/main/java/co/cask/cdap/api/artifact/PluginSelector.java
@@ -18,7 +18,6 @@ package co.cask.cdap.api.artifact;
 
 import co.cask.cdap.api.annotation.Beta;
 import co.cask.cdap.api.templates.plugins.PluginClass;
-import co.cask.cdap.api.templates.plugins.PluginInfo;
 
 import java.util.Map;
 import java.util.SortedMap;
@@ -32,11 +31,11 @@ public class PluginSelector {
   /**
    * Selects a plugin. The default implementation returns the last entry in the given map.
    *
-   * @param plugins set of available plugins. The {@link PluginInfo} is sorted in ascending order of plugin jar
+   * @param plugins set of available plugins. The {@link ArtifactId} is sorted in ascending order of plugin jar
    *                name followed by plugin version.
    * @return a {@link Map.Entry} for the selected plugin.
    */
-  public Map.Entry<ArtifactDescriptor, PluginClass> select(SortedMap<ArtifactDescriptor, PluginClass> plugins) {
+  public Map.Entry<ArtifactId, PluginClass> select(SortedMap<ArtifactId, PluginClass> plugins) {
     return plugins.tailMap(plugins.lastKey()).entrySet().iterator().next();
   }
 }
