+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/runtime/artifact/ArtifactDetail.java
@@ -16,7 +16,7 @@
 
 package co.cask.cdap.internal.app.runtime.artifact;
 
-import co.cask.cdap.api.artifact.ArtifactDescriptor;
+import co.cask.cdap.internal.app.runtime.adapter.ArtifactDescriptor;
 
 /**
  * Details about an artifact, including info about the artifact itself and metadata about the contents of the artifact.
